+++ examples/V7/basicValidation.tsx
@@ -2,12 +2,10 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { useForm } from 'react-hook-form';
 
-export default function App() {
-  const {
-    register,
-    formState: { errors },
-    handleSubmit,
-  } = useForm();
+import './styles.css';
+
+export default function Form() {
+  const { register, errors, handleSubmit } = useForm();
   const onSubmit = (data) => {
     alert(JSON.stringify(data));
   };
@@ -16,22 +14,25 @@ export default function App() {
   return (
     <form onSubmit={handleSubmit(onSubmit)}>
       <div>
-        <label>First name</label>
+        <label htmlFor="firstName">First name</label>
         <input
+          id="firstName"
           type="text"
           {...register('firstName', { required: true, maxLength: 80 })}
         />
       </div>
       <div>
-        <label>Last name</label>
+        <label htmlFor="lastName">Last name</label>
         <input
+          id="lastName"
           type="text"
           {...register('lastName', { required: true, maxLength: 100 })}
         />
       </div>
       <div>
-        <label>Email</label>
+        <label htmlFor="email">Email</label>
         <input
+          id="email"
           type="text"
           {...register('email', {
             required: true,
@@ -41,8 +42,9 @@ export default function App() {
         />
       </div>
       <div>
-        <label>Mobile number</label>
+        <label htmlFor="mobileNumber">Mobile number</label>
         <input
+          id="mobileNumber"
           type="tel"
           {...register('mobileNumber', {
             required: true,
@@ -52,8 +54,8 @@ export default function App() {
         />
       </div>
       <div>
-        <label>Title</label>
-        <select {...register('title', { required: true })}>
+        <label htmlFor="title">Title</label>
+        <select id="title" {...register('title', { required: true })}>
           <option value="Mr">Mr</option>
           <option value="Mrs">Mrs</option>
           <option value="Miss">Miss</option>
@@ -62,21 +64,29 @@ export default function App() {
       </div>
 
       <div>
-        <label>Are you a developer?</label>
-        <input
-          type="radio"
-          value="Yes"
-          {...register('developer', { required: true })}
-        />
-        <input
-          type="radio"
-          value="No"
-          {...register('developer', { required: true })}
-        />
-      </div>
+        <fieldset>
+          <legend>Are you a developer?</legend>
 
-      <div>
-        <input type="text" {...register('asdasd')} />
+          <div>
+            <input
+              id="devYes"
+              type="radio"
+              value="Yes"
+              {...register('developer', { required: true })}
+            />
+            <label htmlFor="devYes">Yes</label>
+          </div>
+
+          <div>
+            <input
+              id="devNo"
+              type="radio"
+              value="No"
+              {...register('developer', { required: true })}
+            />
+            <label htmlFor="devNo">No</label>
+          </div>
+        </fieldset>
       </div>
 
       <input type="submit" />