+++ src/utils/isPlainObject.ts
@@ -1,0 +1,10 @@
+import isObject from './isObject';
+
+export default (tempObject: object) => {
+  const prototypeCopy =
+    tempObject.constructor && tempObject.constructor.prototype;
+
+  return (
+    isObject(prototypeCopy) && prototypeCopy.hasOwnProperty('isPrototypeOf')
+  );
+};